remove old combobox
