+++ vuu-ui/packages/vuu-filters/src/filter-clause/ExpandoCombobox.css
@@ -1,11 +1,8 @@
 .vuuExpandoCombobox {
   --expando-combobox-height: var(--vuuExpandoCombobox-height, var(--salt-size-base));
   --expando-combobox-fontSize: var(--vuuExpandoCombobox-fontSize, 12px);
-
-  --saltInput-outline: none;
-  --saltInput-fontSize: var(--expando-combobox-fontSize);
-  --saltInput-height: var(--expando-combobox-height);
-  --saltInput-minWidth: 4px;
+  --saltInput-paddingLeft: 0px;
+  --saltInput-paddingRight: 0px;
 
   display: inline-flex;
   flex-direction: column;
@@ -13,39 +10,41 @@
   min-width: 4px;
   padding: 0;
 
+  .saltPill {
+    height: 100%;
+    padding-block: 0px;
+  }
+
+  .saltPillInput {
+    min-height: var(--expando-combobox-height);
+    min-width: 0px;
+  }
+
+
+  .saltPillInput-inputWrapper {
+    height: 100%;
+    min-width: max(var(--salt-spacing-800), 100%);
+    padding: 0;
+  }
+  .saltPillInput-input {
+    width: 0;
+  }
+
+  .saltPillInput-endAdornmentContainer {
+    display: none;
+  }
+  .saltPillInput-activationIndicator {
+    display: none;
+  }
+
 }
 
-.vuuExpandoCombobox .saltInput {
-  background-color: transparent;
-  position: absolute;
+.vuuExpandoOption {
+  min-width: calc(var(--salt-size-base) * 2);
+  padding-left: var(--salt-spacing-200);
+  padding-right: var(--salt-spacing-200);
 }
 
-.vuuExpandoCombobox .vuuDropdown {
-  height: 100%;
-}
-
-/** double up the selector just to increase specificity, won't need when we use cascade layers */
-.vuuExpandoCombobox-Input.saltInput {
-  border: none;
-  left: 0px;
-  margin: 0;
-  min-height: 100%;
-  padding: 0;
-  right: 0px;
-  width: auto;
-}
-
-.vuuExpandoCombobox .saltInput-input {
-  border: none;
-  box-sizing: content-box;
-  display: block;
-  flex: 1;
-  font: inherit;
-  margin: 0;
-  min-width: 0;
-  outline: none;
-  padding: 0;
-}
 
 .vuuExpandoCombobox:before {
   content: attr(data-text);
@@ -53,7 +52,5 @@
   font-size: var(--expando-combobox-fontSize);
   height: 0px;
   overflow: hidden;
-  /* visibility: hidden; */
   white-space: pre-wrap;
-  /* position: absolute; */
 }
