+++ vuu-ui/packages/vuu-filters/src/filter-clause/index.ts
@@ -1,4 +1,3 @@
 export * from "./ExpandoCombobox";
-export * from "./ExpandoComboboxSalt";
 export * from "./FilterClause";
 export * from "./value-editors/FilterClauseValueEditorText";
