+++ vuu-ui/packages/vuu-ui-controls/src/index.ts
@@ -1,17 +1,15 @@
 export * from "./calendar";
-export * from "./combo-box";
 export * from "./common-hooks";
 export * from "./cycle-state-button";
 export * from "./date-input";
 export * from "./drag-drop";
 export * from "./date-popup";
 export * from "./drag-drop";
-export * from "./dropdown";
+export * from "./dropdown-base";
 export * from "./editable";
 export * from "./editable-label";
 export * from "./expando-input";
 export * from "./icon-button";
-export * from "./inputs";
 export * from "./instrument-picker";
 export * from "./instrument-search";
 export * from "./list";
