--- vuu-ui/packages/vuu-ui-controls/src/inputs/Checkbox.css
@@ -1,13 +1,0 @@
-.vuuCheckbox {
-  --vuuCheckboxIcon-background-checked-enabled: var(--vuu-color-purple-10);
-  display: flex;
-  height: 24px;
-  align-items: center;
-  gap: 6px;
-  color: var(--light-text-primary, #15171B);
-  font-family: Nunito Sans A-Variant, serif;
-  font-feature-settings: 'ss02' on, 'ss01' on, 'salt' on, 'liga' off;
-  font-size: 12px;
-  font-weight: 400;
-  cursor: pointer;
-}
