--- vuu-ui/packages/vuu-ui-controls/src/inputs/RadioButton.css
@@ -1,27 +1,0 @@
-.vuuRadioButton {
-  display: flex;
-  height: 24px;
-  align-items: center;
-  gap: 6px;
-  color: var(--light-text-primary, #15171B);
-  font-family: Nunito Sans A-Variant, sans-serif;
-  font-feature-settings: 'ss02' on, 'ss01' on, 'salt' on, 'liga' off;
-  font-size: 12px;
-  font-weight: 400;
-  cursor: pointer;
-}
-
-.radio {
-  position: relative;
-  height: 12px;
-}
-
-input[type="radio"] {
-	position: absolute;
-  appearance: none;
-	width: 100%;
-  height: 100%;
-  border-radius: 50%;
-  margin: 0;
-  top: 0;
-}
