--- vuu-ui/packages/vuu-ui-controls/src/inputs/index.ts
@@ -1,2 +1,0 @@
-export * from "./Checkbox";
-export * from "./RadioButton";
