+++ vuu-ui/showcase/src/examples/Layout/index.ts
@@ -11,5 +11,4 @@ export * as LayoutProvider from "./LayoutProvider.examples";
 export * as Stack from "./Stack.examples";
 export * as Toolbar from "./Toolbar.examples";
 export * as View from "./View.examples";
-export * as SavePanel from "./SavePanel.examples";
 export * as LayoutsBrowser from "./LayoutsBrowser.examples";
