+++ vuu-ui/showcase/src/examples/Shell/index.ts
@@ -4,6 +4,7 @@ export * as ConnectionMetrics from "./ConnectionMetrics.examples";
 export * as Feature from "./Feature.examples";
 export * as LeftNav from "./LeftNav.examples";
 export * as LoginPanel from "./LoginPanel.examples";
+export * as SavePanel from "./SavePanel.examples";
 export * as SessionTableEditing from "./SessionTableEditing.examples";
 export * as Shell from "./Shell.examples";
 export * as ThemeProvider from "./ThemeProvider.examples";
